pyproject.toml
[tool.poetry]
name = "deals-project"
version = "0.1.0"
description = ""
authors = ["Raneemoqaily7 <raneemoqaily7@gmail.com>"]
readme = "README.md"
packages = [{include = "deals_project"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.1"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"
django-datetime-utc = "^1.0.4"
django-phonenumber-field = "^7.1.0"
phonenumbers = "^8.13.11"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
